WIP: [Frontend] Create dashboard page
