+++ resources/js/App.tsx
@@ -1,0 +1,7 @@
+import View from "./pages/layouts/View";
+
+function App() {
+  return <View />;
+}
+
+export default App;
