--- resources/js/pages/App.tsx
@@ -1,7 +1,0 @@
-import View from "./View";
-
-function App() {
-  return <View />;
-}
-
-export default App;
