+++ resources/js/pages/dashboard/Activities.tsx
@@ -1,0 +1,16 @@
+import { Box } from "@chakra-ui/react";
+import Activity from "./Activity";
+
+export default function Activities() {
+  return (
+    <Box width="container.sm">
+      <Activity />
+      <Activity />
+      <Activity />
+      <Activity />
+      <Activity />
+      <Activity />
+      <Activity />
+    </Box>
+  );
+}
