--- resources/js/pages/dashboard/Dashboard.tsx
@@ -1,11 +1,0 @@
-import { Fragment } from "react";
-
-import Navbar from "./components/Navbar";
-
-export default function Dashboard() {
-  return (
-    <Fragment>
-      <Navbar />
-    </Fragment>
-  );
-}
