+++ resources/js/pages/dashboard/index.tsx
@@ -1,0 +1,10 @@
+import Container from "../layouts/components/Container";
+import Activities from "./Activities";
+
+export default function Dashboard() {
+  return (
+    <Container>
+      <Activities />
+    </Container>
+  );
+}
