+++ resources/js/pages/follows/index.tsx
@@ -1,0 +1,10 @@
+import { Heading } from "@chakra-ui/react";
+import Container from "../layouts/components/Container";
+
+export default function Follows() {
+  return (
+    <Container>
+      <Heading>Follows</Heading>
+    </Container>
+  );
+}
