+++ resources/js/pages/layouts/Layout.tsx
@@ -1,0 +1,12 @@
+import { Fragment } from "react";
+import { Outlet } from "react-router-dom";
+import Navbar from "./components/Navbar";
+
+export default function Layout() {
+  return (
+    <Fragment>
+      <Navbar />
+      <Outlet />
+    </Fragment>
+  );
+}
