+++ resources/js/pages/layouts/View.tsx
@@ -1,0 +1,13 @@
+import { BrowserRouter } from "react-router-dom";
+
+import ViewRoutes from "./ViewRoutes";
+
+function View() {
+  return (
+    <BrowserRouter>
+      <ViewRoutes />
+    </BrowserRouter>
+  );
+}
+
+export default View;
