+++ resources/js/pages/lessons/index.tsx
@@ -1,0 +1,10 @@
+import { Heading } from "@chakra-ui/react";
+import Container from "../layouts/components/Container";
+
+export default function Lessons() {
+  return (
+    <Container>
+      <Heading>Lessons</Heading>
+    </Container>
+  );
+}
