+++ resources/js/themes/index.ts
@@ -2,8 +2,8 @@ import { extendTheme } from "@chakra-ui/react";
 
 const theme = extendTheme({
   config: {
-    useSystemColorMode: true,
-    initialColorMode: "dark",
+    initialColorMode: "light",
+    useSystemColorMode: false,
   },
   fonts: {
     heading: "Nunito, sans-serif",
